requirements.txt
# 데이터베이스
mysql-connector-python==8.0.33
influxdb-client==1.38.0

# 데이터 처리
pandas==2.0.3
numpy==1.24.3

# 머신러닝
scikit-learn==1.3.0
joblib==1.3.1

# 스케줄링
schedule==1.2.0

# 유틸리티
python-dotenv==1.0.0
psutil==5.9.5
pyyaml==6.0.1

# 로깅 및 모니터링
python-json-logger==2.0.7

# 시각화 (선택적)
matplotlib==3.7.2
seaborn==0.12.2

# 테스트 (개발용)
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1